+++ actions/deploy.go
@@ -7,78 +7,8 @@ import (
 
 	"github.com/CenturyLinkLabs/prettycli"
 	"github.com/CenturyLinkLabs/zodiac/cluster"
-	"github.com/CenturyLinkLabs/zodiac/composer"
-	"github.com/CenturyLinkLabs/zodiac/proxy"
 	"github.com/samalba/dockerclient"
 )
-
-const ProxyAddress = "localhost:31981"
-
-var (
-	DefaultProxy    proxy.Proxy
-	DefaultComposer composer.Composer
-)
-
-func init() {
-	DefaultProxy = proxy.NewHTTPProxy(ProxyAddress)
-	DefaultComposer = composer.NewExecComposer(ProxyAddress)
-}
-
-type DeploymentManifests []DeploymentManifest
-
-type DeploymentManifest struct {
-	Services   []Service
-	DeployedAt string
-}
-
-type Service struct {
-	Name            string
-	ContainerConfig dockerclient.ContainerConfig
-}
-
-func Rollback(c cluster.Cluster, args []string) (prettycli.Output, error) {
-	var reqs []cluster.ContainerRequest
-
-	for _, endpoint := range c.Endpoints() {
-
-		reqs = collectRequests(endpoint, args)
-		// Get most recent deployment's manifests
-		var manifests DeploymentManifests
-		for _, req := range reqs {
-			ci, err := endpoint.InspectContainer(req.Name)
-			if err != nil {
-				continue
-			}
-
-			if err := json.Unmarshal([]byte(ci.Config.Labels["zodiacManifest"]), &manifests); err != nil {
-				fmt.Printf("ERROR: %s\n\n", err)
-				return nil, err
-			}
-
-			break
-		}
-
-		// shut down current deployment
-
-		currentDeployment := manifests[len(manifests)-1]
-
-		for _, svc := range currentDeployment.Services {
-			endpoint.RemoveContainer(svc.Name)
-		}
-
-		// TODO: allow passing in index
-		newDeployment := manifests[len(manifests)-2]
-
-		manifests = append(manifests, newDeployment)
-		newDeployment = manifests[len(manifests)-1]
-		manifests[len(manifests)-1].DeployedAt = time.Now().Format(time.RFC3339)
-
-		startServices(newDeployment.Services, manifests, endpoint)
-	}
-
-	output := fmt.Sprintf("Successfully rolled back to %d container(s)", len(reqs))
-	return prettycli.PlainOutput{output}, nil
-}
 
 func Deploy(c cluster.Cluster, args []string) (prettycli.Output, error) {
 
@@ -151,34 +81,4 @@ func serviceForRequest(req cluster.ContainerRequest) (Service, error) {
 		Name:            req.Name,
 		ContainerConfig: cc,
 	}, nil
-}
-
-func collectRequests(endpoint cluster.Endpoint, args []string) []cluster.ContainerRequest {
-	// TODO: handle error
-	go DefaultProxy.Serve(endpoint)
-	// TODO: handle error
-	defer DefaultProxy.Stop()
-
-	// TODO: handle error
-	// TODO: args not passed!
-	DefaultComposer.Run(args)
-	return DefaultProxy.DrainRequests()
-}
-
-func startServices(services []Service, manifests DeploymentManifests, endpoint cluster.Endpoint) error {
-	manifestsBlob, err := json.Marshal(manifests)
-	if err != nil {
-		return err
-	}
-
-	for _, svc := range services {
-		if svc.ContainerConfig.Labels == nil {
-			svc.ContainerConfig.Labels = make(map[string]string)
-		}
-		svc.ContainerConfig.Labels["zodiacManifest"] = string(manifestsBlob)
-
-		endpoint.StartContainer(svc.Name, svc.ContainerConfig)
-	}
-
-	return nil
 }
